FlipTrack/pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask-login>=0.6.3",
    "flask>=3.1.2",
    "flask-sqlalchemy>=3.1.1",
    "flask-wtf>=1.2.1",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "trafilatura>=2.0.0",
    "pillow>=11.3.0",
    "beautifulsoup4>=4.13.4",
    "requests>=2.32.5",
    "pillow-heif>=1.1.0",
    "sqlalchemy>=2.0.43",
    "werkzeug>=3.1.3",
]
